refactor: rename qr code scanner package
