+++ presentation/src/main/java/com/yapp/presentation/ui/AttendanceScreen.kt
@@ -12,7 +12,7 @@ import com.yapp.common.util.KakaoTalkLoginProvider
 import com.yapp.presentation.ui.admin.main.AdminMain
 import com.yapp.presentation.ui.login.Login
 import com.yapp.presentation.ui.member.main.MemberMain
-import com.yapp.presentation.ui.member.qrcodescan.QrCodeScanner
+import com.yapp.presentation.ui.member.qrcodescanner.QrCodeScanner
 import com.yapp.presentation.ui.member.setting.MemberSetting
 import com.yapp.presentation.ui.splash.Splash
